Simplify EdgeMap multiplication to only one implementation.
The pure bit-twiddling implementation, was complex and unsafe but faster.
However, using theses in-practice requires an branch in-direction. When
the branch is predicted 1/3 cycles vs a multiply depending on the CPU. But,
if the turns are random, and branch prediction fails, they are marginally
slower.

Previously, `CornerMap` had a similar bit-twiddling implementation and when
only in-direction was used to switch for both the `CornerMap` & `EdgeMap` in
Cube`, the performance was universally better. However, the representation
of the orientation for the corners was flawed, as orientation of corners
relative to current positions (meaning the index's where not independent).
Fixing the representation issue of `CornerMap` made the direct bit-twiddling
approach less-efficient.

For now, the simplified, multiply-only, implementation will be used.