Clearly separate validity from solve-ability
This allows for well-defined behaviour for representing mapping which
are outside the standard configurations space of the cube (that is
unsolvable) but which still define a valid permutation with all values
in bounds.

Additionally, we can have stronger guarantees as without using unsafe
the api should not allow you to construct an invalid mapping. However,
since individual components (ex Corners or Edges) can be combined
independently it easy to create a valid by unsolvable cube.

Note: that these valid but unsolvable cubes are useful.